requirements.txt
pytest==8.3.5
pytest-bdd==7.3.0
playwright==1.52.0
python-dotenv==1.0.1